fix(server): respect basePath in swagger api
